rag_clients/vertex_rag.py
"""
Vertex AI RAG Client - Wrapper for Vertex AI RAG-managed vector stores
Handles retrieval from configured RAG engines/corpora
"""
from typing import Dict, List, Any, Optional
import asyncio
from datetime import datetime
from google.cloud import aiplatform
from vertexai.preview import rag
import vertexai

from utils.logging import get_logger

logger = get_logger()


class VertexRAGClient:
    """Client for Vertex AI RAG Engine retrieval"""
    
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.project_id = config["project"]["gcp_project_id"]
        self.location = config["project"]["location"]
        
        # Initialize Vertex AI
        vertexai.init(project=self.project_id, location=self.location)
        aiplatform.init(project=self.project_id, location=self.location)
        
        logger.info(
            f"Initialized Vertex RAG client: "
            f"project={self.project_id}, location={self.location}"
        )
    
    async def search(
        self,
        engine_name: str,
        query: str,
        config: Dict[str, Any]
    ) -> Dict[str, Any]:
        """
        Search a specific RAG engine
        
        Args:
            engine_name: Name of the engine (e.g., 'legal', 'gos')
            query: Search query
            config: Engine-specific configuration with filters and top_k
        
        Returns:
            Dictionary with documents and metadata
        """
        start_time = datetime.utcnow()
        
        try:
            engine_config = self.config["engines"][engine_name]
            rag_corpus_id = engine_config["id"]
            top_k = config.get("top_k", 20)
            
            logger.info(
                f"Searching engine '{engine_name}': "
                f"query='{query[:50]}...', top_k={top_k}"
            )
            
            # Build filter from config
            filter_dict = self._build_filter(config.get("filters", {}))
            
            # Execute RAG retrieval
            # Note: This is a simplified example. Adjust based on actual Vertex RAG API
            response = await self._execute_rag_query(
                rag_corpus_id=rag_corpus_id,
                query=query,
                top_k=top_k,
                filter_dict=filter_dict
            )
            
            # Parse response
            documents = self._parse_response(response, engine_name)
            
            end_time = datetime.utcnow()
            latency_ms = int((end_time - start_time).total_seconds() * 1000)
            
            logger.info(
                f"Engine '{engine_name}' returned {len(documents)} documents "
                f"in {latency_ms}ms"
            )
            
            return {
                "engine": engine_name,
                "documents": documents,
                "count": len(documents),
                "latency_ms": latency_ms,
                "status": "success"
            }
            
        except Exception as e:
            logger.error(f"Search failed for engine '{engine_name}': {e}", exc_info=True)
            return {
                "engine": engine_name,
                "documents": [],
                "count": 0,
                "latency_ms": 0,
                "status": "error",
                "error": str(e)
            }
    
    async def _execute_rag_query(
        self,
        rag_corpus_id: str,
        query: str,
        top_k: int,
        filter_dict: Dict[str, Any]
    ) -> Any:
        """
        Execute the actual RAG query using Vertex AI SDK
        
        Note: This is a simplified implementation. Adjust based on your
        actual Vertex RAG API version and methods.
        """
        try:
            # Use the RAG API to retrieve contexts
            # Adjust this based on actual Vertex RAG API
            
            # Example using rag.retrieval.retrieval
            # response = rag.retrieval.retrieve(
            #     rag_corpus=rag_corpus_id,
            #     query=query,
            #     top_k=top_k,
            #     filter=filter_dict
            # )
            
            # For now, return mock response structure
            # REPLACE THIS with actual API call
            response = {
                "contexts": [
                    {
                        "source_uri": f"gs://bucket/doc_{i}.pdf",
                        "text": f"Mock content {i} for query: {query}",
                        "score": 0.9 - (i * 0.05),
                        "metadata": {
                            "source": "mock_source",
                            "page": i + 1
                        }
                    }
                    for i in range(min(top_k, 5))  # Mock 5 results
                ]
            }
            
            return response
            
        except Exception as e:
            logger.error(f"RAG query execution failed: {e}")
            raise
    
    def _build_filter(self, filters: Dict[str, Any]) -> Dict[str, Any]:
        """
        Build Vertex RAG filter dictionary from config filters
        
        Adjust based on your metadata structure in Vertex RAG
        """
        filter_dict = {}
        
        if filters.get("jurisdiction"):
            filter_dict["jurisdiction"] = filters["jurisdiction"]
        
        if filters.get("years"):
            filter_dict["year"] = {"in": filters["years"]}
        
        if filters.get("districts"):
            filter_dict["district"] = {"in": filters["districts"]}
        
        if filters.get("go_numbers"):
            filter_dict["go_number"] = {"in": filters["go_numbers"]}
        
        # Add more filter mappings as needed
        
        return filter_dict
    
    def _parse_response(
        self,
        response: Any,
        engine_name: str
    ) -> List[Dict[str, Any]]:
        """Parse Vertex RAG response into standardized document format"""
        documents = []
        
        # Adjust based on actual response structure
        contexts = response.get("contexts", [])
        
        for i, context in enumerate(contexts):
            doc = {
                "id": f"{engine_name}_{i}",
                "vertical": engine_name,
                "source_uri": context.get("source_uri", ""),
                "text": context.get("text", ""),
                "score": context.get("score", 0.0),
                "metadata": context.get("metadata", {}),
                "rank": i
            }
            
            # Extract locator (page, section, etc.) from metadata
            metadata = context.get("metadata", {})
            locator = metadata.get("page", metadata.get("section", ""))
            doc["locator"] = str(locator) if locator else ""
            
            # Extract source date if available
            doc["source_date"] = metadata.get("date", metadata.get("published_date", ""))
            
            documents.append(doc)
        
        return documents


# Async helper for batch retrieval
async def batch_retrieve(
    client: VertexRAGClient,
    queries: List[Dict[str, Any]]
) -> List[Dict[str, Any]]:
    """
    Execute multiple RAG queries in parallel
    
    Args:
        client: VertexRAGClient instance
        queries: List of query dictionaries with engine_name, query, config
    
    Returns:
        List of results
    """
    tasks = [
        client.search(
            engine_name=q["engine_name"],
            query=q["query"],
            config=q["config"]
        )
        for q in queries
    ]
    
    results = await asyncio.gather(*tasks, return_exceptions=True)
    
    return [r if not isinstance(r, Exception) else {"status": "error", "error": str(r)} 
            for r in results]


if __name__ == "__main__":
    # Test the RAG client
    import asyncio
    from config import load_config
    
    config = load_config()
    
    async def test():
        client = VertexRAGClient(config)
        
        test_config = {
            "top_k": 10,
            "filters": {"jurisdiction": "Andhra Pradesh"}
        }
        
        result = await client.search(
            engine_name="legal",
            query="teacher transfer rules",
            config=test_config
        )
        
        print(f"Status: {result['status']}")
        print(f"Documents: {result['count']}")
        if result['documents']:
            print(f"First doc: {result['documents'][0]['text'][:100]}")
    
    asyncio.run(test())