pyproject.toml
[project]
name = "ai-for-ap-school-education"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.9.1",
    "asyncio==3.4.3",
    "black==24.1.1",
    "dateparser==1.2.0",
    "fastapi==0.109.0",
    "google-cloud-aiplatform>=1.47.0",
    "google-cloud-storage>=2.18.0",
    "httpx>=0.27.0",
    "langchain>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-google-vertexai>=1.1.0",
    "langgraph>=0.2.0",
    "numpy==1.26.3",
    "pandas==2.2.0",
    "prometheus-client==0.19.0",
    "pydantic>=2.9",
    "pydantic-settings>=2.4.0",
    "pytest==8.0.0",
    "pytest-asyncio==0.23.3",
    "python-dateutil==2.8.2",
    "python-json-logger==2.0.7",
    "python-multipart==0.0.6",
    "pyyaml==6.0.1",
    "ruff==0.1.14",
    "sentence-transformers==2.3.1",
    "spacy==3.7.2",
    "structlog==24.1.0",
    "uvicorn[standard]==0.27.0",
    "vertexai>=1.47.0",
]
